Consolidate BufferAccumulator to buffer-api module
