+++ data-prepper-plugins/buffer-common/src/test/java/org/opensearch/dataprepper/buffer/common/BufferAccumulatorTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.source.opensearch.worker;
+package org.opensearch.dataprepper.buffer.common;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
