+++ data-prepper-plugins/s3-source/build.gradle
@@ -13,6 +13,7 @@ repositories {
 
 dependencies {
     implementation project(':data-prepper-api')
+    implementation project(':data-prepper-plugins:buffer-common')
     implementation project(':data-prepper-plugins:common')
     implementation libs.armeria.core
     implementation 'io.micrometer:micrometer-core'
