+++ settings.gradle
@@ -121,3 +121,4 @@ include 'data-prepper-plugins:aws-plugin-api'
 include 'data-prepper-plugins:aws-plugin'
 include 'data-prepper-plugins:obfuscate-processor'
 include 'data-prepper-plugins:parquet-codecs'
+include 'data-prepper-plugins:buffer-common'
